Finish STORE operations
Finishes the STORE operations:
* STJ - Store the Jump register
        This differs from STi operations in that
	the default F-modification is 2 (0:2).
	This is because the Jump usually stores an address,
	and (0:2) is the address part of a word used as an instruction.
* STZ - Like STA, but uses +|0|0|0|0|0 instead of the Accumulator.

Also fixed a bug in set-data-sign, which was incorrectly trying to use
assoc-in, rather than assoc.

And fixed another bug the field specification STORE application
function, which was incorrectly defaulting to :plus if the sign was not
part of the field (should have chosen the destination's sign instead).