Address transfer operations
All address transfer operations implemented:
* ENTA, ENTX, ENTi (for i [1, 6])
* ENNA, ENNX, ENNi (for i [1, 6])
* INCA, INCX, INCi (for i [1, 6])
* DECA, DECX, DECi (for i [1, 6])

Additional required changes:
* Refactored addition and subtraction, to pull out the basic operation
of MIX machine addition, which is a function of the machine, 2 operands,
and a destination register.
* Added a function to add two MIX data elements together, returning a
MIX data element (rather than an integer).  If the result is 0, the sign
is set to the sign of the first operand.
* Changed address indexing function to return M as a MIX data word,
rather than an integer.  This was needed for ENT* operations, since
there is the requirement that, if M = 0, the sign of the instruction is
used as the resulting sign.
* All operations needed to be updated to handle M as a MIX data word.