Jump operations
Completes all jump operations.
This required a few changes to the machine, and how programs are
executed:
* The machine needed some way of keeping track of which instruction is
the current / next instruction, so a program counter was added.  This
program counter is used to fetch the next instruction, and is
incremented at the end of each instruction.  For jump operations (except
JSJ) the program counter is incremented and stored in rJ *before* the
program counter is set to M.
* Added a means of loading a program into memory.  For now, programs are
loaded started at memory 0.  The program counter is updated to the start
of the program.
* execute-program now only takes a machine, and assumes the program has
already been loaded, and the program counter points to the first
instruction.  It then begins executing instructions until a HLT
instruction is encountered, and returns the machine.