Better type-safety for data namespace
This commit incorporates clojure.spec.alpha to provide better type
checking for the data namespace.  This is to facilitate more complete
tests of this namespace.

Defines specs for:
* sign
* byte
* bytes
* data (note - this is where it's too restrictive)

Adds pre and post conditions to all functions.  This makes it more
strict, in that those functions that don't necessarily need complete MIX
data now require it (i.e. set-sign really only needs the sign, but now
requires a complete MIX 'word').  The restriction may or may not be a
benefit - it is yet to be seen.