Finish data tests
Adds tests for the rest of data namespace.  Also makes some changes to
data namespace, either in response to testing (i.e. bug fixes etc.) or
to make testing easier / more thorough:

* Changed split size argument to be constrained to a positive integer,
instead of the range [1, (count-bytes data)].  In the case that size is
>= to (count-bytes data), a vector containing data is returned.

* Changed merge to use concat rather than into, since into does not
accept variable args like concat does, and the result is good enough
(concat returns a list, but new-data will convert it to a vector
anyway).

* Changed cycle-* functions to return data if it has no bytes (i.e. an
empty vector of bytes).  This is because cycle needs to check for n >
(count-bytes data), and mod to get the correct shift amount.  If
(count-bytes data) = 0, then (mod n 0) throws an exception (divide by
zero).

* Added spec for a character sequence, along with valid-char-seq? and
assert-char-seq functions.