Field spec tests
Adds tests for the field-spec namespace, covering all functions.

* Adds a specification for a field spec, to be used with the spec
namespace functions for validations and explanations of field specs.

* Adds argument validation for all field-spec functions.

* Adds a definition of a MIX word-size to data namespace.

* Adds an assertion that some MIX data is a valid MIX word (i.e.
validates the number of bytes, as well as the data).  This was added to
field-spec namespace, since it is currently the only namespace that
needs it.  In the future, it can easily be moved to data namespace.