Machine namespace tests
* Fixed a bug in get/set-index-register, which would allow arbitrary
register formats, so long as it was a sequence of at least 2 things,
where the second thing was an integer between [1, 6], it would allow it.
It now requires the format [:I i], where i is an int in range [1, 6].

* Changed the error handling of set-register to assert rather than
explicitly throw.  This fits the model other namespaces are using (e.g.
data namespace), and more closely fits our use case, which is to ensure
the arguments to the function are valid.  I believe the language
typically uses asserts for this sort of thing as well.

* Added validation of register arguments to functions (incomplete).
* Added validation of MIX data, when appropriate (incomplete)

* Removed magic numbers from machine namespace.  There are now
  definitions for:
  - word: number of bytes for a MIX word
  - half-word: number of bytes for a MIX 'half-word'
  - mem-limit: amount of available memory, in words
  - tape-ports: number of tape devices that can be attached to machine
  - disk-ports: number of disk devices that can be attached to machine

* Added validation and assertions for memory locations and ranges
  for both get-memory and set-memory.  This was done by validating both
  the starting memory location, the size argument, and the resulting
  range.  Additionally, the size argument is validated as a positive
  integer.

* Added memory location validation to set-program-counter, and
inc-program-counter.